pyproject.toml
[tool.poetry]
name = "hass-time-machine"
version = "1.0.0"
description = ""
authors = ["gwisp2 <60121533+gwisp2@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
homeassistant = "^2022.8.3"

[tool.poetry.dev-dependencies]
mypy = "^0.971"
black = "^22.6.0"
isort = "^5.10.1"
types-requests = "^2.28.8"
flake8 = "^5.0.4"
pylint = "^2.14.5"
voluptuous-stubs = "^0.1.1"
homeassistant-stubs = "^2022.8.6"

[tool.isort]
profile = "black"

[tool.poe.tasks]
isort = "isort custom_components/time_machine"
black = "black custom_components/time_machine"
format = ["isort", "black"]
mypy = "mypy --strict --python-executable=./.venv/bin/python custom_components/time_machine"
flake8 = "flake8 custom_components/time_machine"
pylint = "pylint custom_components/time_machine"
lint = ["mypy", "flake8", "pylint"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
